Fix availability fraction bool arithmetic on analysis datasets
On an analysis dataset the point series has no non-time dims, so the
null-fraction mean over an empty dim list is an identity op that keeps
the mask's bool dtype. The two-point fraction combine then does bool
addition (logical or), rendering fully-missing positions at fraction
0.5 — the brown midpoint of the availability colormap — instead of 0.0
light red, and hiding the distinction from half-missing positions.
Forecast datasets were unaffected (their mean over lead_time produces
floats), which is why existing tests passed.

+++ src/scripts/validation/availability.py
@@ -236,7 +236,11 @@ def _compute_nulls_for_point(
         check_mask = null_mask.isel(lead_time=slice(1, None))
 
     time_dim = next(d for d in ("time", "init_time") if d in da_point.dims)
-    null_frac = check_mask.mean(dim=[d for d in non_time_dims if d in check_mask.dims])
+    # On an analysis dataset the point series has no non-time dims, and mean over an
+    # empty dim list keeps the mask's bool dtype; cast so callers can do arithmetic.
+    null_frac = check_mask.mean(
+        dim=[d for d in non_time_dims if d in check_mask.dims]
+    ).astype(np.float64)
 
     if check_mask.any():
         unavailable = null_frac[time_dim].where(null_frac > 0, drop=True)
